Added support for multi-chip configurations.
This is based on 14-bit ids that consists of both a chip and a tile id.
We represent these ids with the TileId type to keep the exact encoding
at one place. However, it is currently only supported with gem5.

We also have one example configuration (config/multichip.py) that
creates two chips connected via bridges. Therefore, we can also
simulate the increased latency for cross-chip communication.

+++ src/libs/rust/thread/src/lib.rs
@@ -193,7 +193,7 @@ impl Thread {
             // safety: has_msg is true and we trust the TCU
             unsafe {
                 let head = self.msg.as_ptr() as *const tcu::Header;
-                let slice = [head as usize, (*head).length as usize];
+                let slice = [head as usize, (*head).length()];
                 Some(transmute(slice))
             }
         }
@@ -218,7 +218,7 @@ impl Thread {
     }
 
     fn set_msg(&mut self, msg: &'static tcu::Message) {
-        let size = msg.header.length as usize + mem::size_of::<tcu::Header>();
+        let size = msg.header.length() + mem::size_of::<tcu::Header>();
         self.has_msg = true;
         // safety: we trust the TCU
         unsafe {
